we now have 2 different API approaches: CustomRegion_DelegateApproach & CustomRegion_InterfaceApproach.
Let's see what is easier to implement. The API surface actually doesn't feel that different as you can see in the VL patches PatchedFunctionDemo_DelegateApproach & PatchedFunctionDemo_InterfaceApproach. See GPUFunction patches

the main difference is:
* in the interface based approach the compiler has to emit two types that implement certain interfaces that together reflect how the user used your custom region
* in the delegate approach the compiler only to instanciate an already defined type CustomRegion (so no need to emit new types), and then build 2 or 3 delegates and set them on the respective properties on that CustomRegion instance.

the compiler currently already knows how to build delegates that reflect stateful regions. So we'll probably go for the delegate based approach in order to not reinvent too much